move functions into common for better reusability
